readme: remove WIP mention
I personally don't work on this project anymore (day job, life, blabla).
I'll keep this project as-is as it might help somebody.

note: I receive regular enquiries about this project, and some people
seem to be working on some virtio gpu support for windows. Please note
the source of truth regarding this project should be found on
virglrenderer-devel mailing list.

Signed-off-by: Nathan Gauër <nathan@gauer.org>